Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -6,7 +6,7 @@ int main(){
     int sum = a+b;
     int sub = a-b;
     int mul = a*b;
-    int div = a/b;
+    float div = a/b;
     if(ch == '+'){
         printf("%d",sum);
     }
